requirements.txt
flask>=2.0.0
flask-sqlalchemy
flask-migrate
werkzeug
flask-cors
pyjwt
python-dotenv
